+++ packages/element/src/test/create-interpolate-element.tsx
@@ -73,12 +73,19 @@ describe( 'createInterpolateElement', () => {
 			createInterpolateElement( testString, { 'spaced token': <em /> } )
 		).toEqual( expectedElement );
 	} );
-	it( 'returns same string when there is an unmatched closing tag', () => {
+	it( 'returns same string and warns when there is an unmatched closing tag', () => {
 		const testString = 'This is a </item> string';
 		const expectedElement = <>{ testString }</>;
+		const consoleWarn = jest
+			.spyOn( console, 'warn' )
+			.mockImplementation( () => {} );
 		expect(
 			createInterpolateElement( testString, { item: <em /> } )
 		).toEqual( expectedElement );
+		expect( consoleWarn ).toHaveBeenCalledWith(
+			expect.stringContaining( "Unmatched closing tag '</item>'" )
+		);
+		consoleWarn.mockRestore();
 	} );
 	it( 'returns expected react element for non nested components', () => {
 		const testString = 'This is a string with <a>a link</a>.';